Moved error handling, code polish
