+++ src/main/java/net/engio/mbassy/bus/AbstractSyncAsyncMessageBus.java
@@ -8,7 +8,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.TimeUnit;
 
-import net.engio.mbassy.bus.common.IMessageBus;
+import net.engio.mbassy.IMessageBus;
 import net.engio.mbassy.bus.config.Feature;
 import net.engio.mbassy.bus.config.IBusConfiguration;
 import net.engio.mbassy.bus.error.PublicationError;
