+++ src/main/java/net/engio/mbassy/bus/MBassador.java
@@ -2,7 +2,7 @@ package net.engio.mbassy.bus;
 
 import java.util.concurrent.TimeUnit;
 
-import net.engio.mbassy.bus.common.IMessageBus;
+import net.engio.mbassy.IMessageBus;
 import net.engio.mbassy.bus.config.BusConfiguration;
 import net.engio.mbassy.bus.config.Feature;
 import net.engio.mbassy.bus.config.IBusConfiguration;
