+++ src/test/java/net/engio/mbassy/SyncBusTest.java
@@ -2,9 +2,7 @@ package net.engio.mbassy;
 
 import java.util.concurrent.atomic.AtomicInteger;
 
-import net.engio.mbassy._misc.BusFactory;
 import net.engio.mbassy.bus.MBassador;
-import net.engio.mbassy.bus.common.IMessageBus;
 import net.engio.mbassy.bus.error.IPublicationErrorHandler;
 import net.engio.mbassy.bus.error.PublicationError;
 import net.engio.mbassy.common.ConcurrentExecutor;
